Become sole owner of coercion logic.
