+++ bluesky_live/__init__.py
@@ -2,3 +2,8 @@ from ._version import get_versions
 
 __version__ = get_versions()["version"]
 del get_versions
+
+
+from ._coercion import register
+register()
+del register
