+++ bluesky_live/_utils.py
@@ -143,71 +143,3 @@ def _no_op(doc):
     # Implemented as a function (not a lambda)
     # so that objects that have it are pickle-able.
     return doc
-
-
-def coerce_dask(handler_class, filler_state):
-    # If the handler has its own delayed logic, defer to that.
-    if hasattr(handler_class, "return_type"):
-        if handler_class.return_type["delayed"]:
-            return handler_class
-
-    # Otherwise, provide best-effort dask support by wrapping each datum
-    # payload in dask.array.from_delayed. This means that each datum will be
-    # one dask task---it cannot be rechunked into multiple tasks---but that
-    # may be sufficient for many handlers.
-    class Subclass(handler_class):
-        def __call__(self, *args, **kwargs):
-            descriptor = filler_state.descriptor
-            key = filler_state.key
-            shape = extract_shape(descriptor, key)
-            # there is an un-determined size (-1) in the shape, abandon
-            # lazy as it will not work
-            if any(s <= 0 for s in shape):
-                return dask.array.from_array(super().__call__(*args, **kwargs))
-            else:
-                dtype = extract_dtype(descriptor, key)
-                load_chunk = dask.delayed(super().__call__)(*args, **kwargs)
-                return dask.array.from_delayed(load_chunk, shape=shape, dtype=dtype)
-
-    return Subclass
-
-
-def extract_shape(descriptor, key):
-    """
-    Work around bug in https://github.com/bluesky/ophyd/pull/746
-    """
-    # Ideally this code would just be
-    # descriptor['data_keys'][key]['shape']
-    # but we have to do some heuristics to make up for errors in the reporting.
-
-    # Broken ophyd reports (x, y, 0). We want (num_images, y, x).
-    data_key = descriptor["data_keys"][key]
-    if len(data_key["shape"]) == 3 and data_key["shape"][-1] == 0:
-        object_keys = descriptor.get("object_keys", {})
-        for object_name, data_keys in object_keys.items():
-            if key in data_keys:
-                break
-        else:
-            raise RuntimeError(f"Could not figure out shape of {key}")
-        for k, v in descriptor["configuration"][object_name]["data"].items():
-            if k.endswith("num_images"):
-                num_images = v
-                break
-        else:
-            num_images = -1
-        x, y, _ = data_key["shape"]
-        shape = (num_images, y, x)
-    else:
-        shape = descriptor["data_keys"][key]["shape"]
-    return shape
-
-
-def extract_dtype(descriptor, key):
-    """
-    Work around the fact that we currently report jsonschema data types.
-    """
-    reported = descriptor["data_keys"][key]["dtype"]
-    if reported == "array":
-        return float  # guess!
-    else:
-        return reported
